Add selector
