+++ client/package.json
@@ -19,6 +19,7 @@
     "react-select": "^3.0.8",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
     "uuid": "^3.3.3"
   },
   "scripts": {
