+++ client/src/main/Main.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { withStyles } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
-import { getProjects, addProject, deleteProject, updateFilter, applyFilter } from './actions/projectListActions';
+import { getProjects, addProject, deleteProject, updateFilter } from './actions/projectListActions';
+import { getFilteredProjects } from './selectors/projectListSelector';
 import ToolBar from './components/ToolBar';
 import ProjectList from './components/ProjectList';
 import Grid from '@material-ui/core/Grid';
@@ -21,14 +22,13 @@ class Main extends Component {
   }
 
   render() {
-    const { classes, projectList, addProject, deleteProject, updateFilter, applyFilter } = this.props;
+    const { classes, projectList, addProject, deleteProject, updateFilter } = this.props;
     return (
       <Grid className={classes.container} container spacing={2}>
         <Grid item xs={12} sm={3}>
           <ToolBar
             addProject={addProject}
             updateFilter={updateFilter}
-            applyFilter={applyFilter}
           />
         </Grid>
         <Grid item xs={12} sm={9}>
@@ -48,11 +48,13 @@ Main.propTypes = {
   addProject: PropTypes.func.isRequired,
   deleteProject: PropTypes.func.isRequired,
   updateFilter: PropTypes.func.isRequired,
-  applyFilter: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({
-  projectList: state.projectList,
+  projectList: {
+    ...state.projectList,
+    filteredProjects: getFilteredProjects(state),
+  }
 });
 
 export default connect(
@@ -62,6 +64,5 @@ export default connect(
     addProject,
     deleteProject,
     updateFilter,
-    applyFilter,
   },
 )(withStyles(styles)(Main));
