+++ client/src/main/actions/projectListActions.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { GET_PROJECTS, ADD_PROJECT, DELETE_PROJECT, PROJECTS_LOADING, UPDATE_FILTER, UPDATE_PROJECTS } from './types';
+import { GET_PROJECTS, ADD_PROJECT, DELETE_PROJECT, PROJECTS_LOADING, UPDATE_FILTER } from './types';
 
 export const getProjects = () => dispatch => {
   dispatch(setProjectsLoading());
@@ -8,10 +8,7 @@ export const getProjects = () => dispatch => {
       dispatch({
         type: GET_PROJECTS,
         payload: response.data,
-      });
-      dispatch({
-        type: UPDATE_PROJECTS,
-        payload: response.data,
+        loading: false,
       });
     });
 };
@@ -59,32 +56,6 @@ export const updateFilter = (filterType, data) => dispatch => {
     type: UPDATE_FILTER,
     filterType,
     data,
-  });
-};
-
-export const applyFilter = () => (dispatch, getState) => {
-  const state = getState();
-  const { projects, filters } = state.projectList;
-  const { teamSizeFilter, categoryFilter, technologyFilter } = filters;
-
-  dispatch(setProjectsLoading());
-
-  const filteredProjects = projects.filter(
-    project => {
-      const { size, category: projectCategory, technology: projectTechnology } = project;
-      const noSize = Object.keys(teamSizeFilter).length === 0 && teamSizeFilter.constructor === Object;
-      const inSize = teamSizeFilter.minValue <= size && size <= teamSizeFilter.maxValue;
-      const noCategory = Array.isArray(categoryFilter) && !categoryFilter.length;
-      const inCategory = categoryFilter.every(category => projectCategory.includes(category));
-      const noTechnology = Array.isArray(technologyFilter) && !technologyFilter.length;
-      const inTechnology = technologyFilter.every(technology => projectTechnology.includes(technology));
-      return (noSize || inSize) && (noCategory || inCategory) && (noTechnology || inTechnology);
-    }
-  );
-
-  dispatch({
-    type: UPDATE_PROJECTS,
-    payload: filteredProjects,
   });
 };
