+++ client/src/main/actions/types.js
@@ -3,4 +3,3 @@ export const ADD_PROJECT = 'ADD_PROJECT';
 export const DELETE_PROJECT = 'DELETE_PROJECT';
 export const PROJECTS_LOADING = 'PROJECTS_LOADING';
 export const UPDATE_FILTER = 'UPDATE_FILTER';
-export const UPDATE_PROJECTS = 'UPDATE_PROJECTS';
